requirements.txt
ultralytics
streamlit
streamlit-webrtc
av
opencv-python-headless
pillow
numpy